requirements.txt
# Flask Attendance System - Requirements
# Phase 1: Foundation Setup

# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.4
Flask-CORS==4.0.0
Flask-WTF==1.2.1

# Database
SQLAlchemy==2.0.23
alembic==1.12.1

# Task Queue & Caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Face Recognition & Computer Vision
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
dlib==19.24.2

# Authentication & Security
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
cryptography==41.0.7

# API & Serialization
marshmallow==3.20.1
Flask-Marshmallow==0.15.0

# Real-time Communication
python-socketio==5.10.0
eventlet==0.33.3

# Reporting & Export
openpyxl==3.1.2
WeasyPrint==60.1
reportlab==4.0.7
pandas==2.1.4
matplotlib==3.8.2

# HTTP & Requests
requests==2.31.0
httpx==0.25.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3.post1

# Validation
email-validator==2.1.0
validators==0.22.0

# Utilities
click==8.1.7
python-magic==0.4.27
bleach==6.1.0

# Monitoring & Logging
python-json-logger==2.0.7
sentry-sdk[flask]==1.38.0

# Testing (Development)
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.1.0
factory-boy==3.3.0

# Code Quality (Development)
black==23.12.1
flake8==6.1.0
pylint==3.0.3
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Production Server
gunicorn==21.2.0
gevent==23.9.1

# Environment Management
python-decouple==3.8